moved ColumnInfo class into Diffix project the class was also renamed as DColumnInfo
