+++ src/explorer.api/Exploration/Setup/ContextBuilder.cs
@@ -39,7 +39,7 @@
                 {
                     throw new MetaDataCheckException($"Could not find column '{dataSource}.{table}.{column}'.");
                 }
-                var ci = new ColumnInfo(columnInfo.Type, columnInfo.UserId, columnInfo.Isolating ?? true);
+                var ci = new DColumnInfo(columnInfo.Type, columnInfo.UserId, columnInfo.Isolating ?? true);
                 return new CheckedContext(dataSource, table, column, ci);
             });
         }
@@ -50,7 +50,7 @@
         /// </summary>
         private class CheckedContext : ExplorerContext
         {
-            internal CheckedContext(string dataSource, string table, string column, ColumnInfo columnInfo)
+            internal CheckedContext(string dataSource, string table, string column, DColumnInfo columnInfo)
             {
                 DataSource = dataSource;
                 Table = table;
@@ -64,7 +64,7 @@
 
             public string Column { get; }
 
-            public ColumnInfo ColumnInfo { get; }
+            public DColumnInfo ColumnInfo { get; }
         }
     }
 }
